@TianQiLin666666 2025-05-12
> Nice PR! And thanks for the detailed benchmark!
> 
> How about provide two optional server args:
> 
> * prefill attention backend
> * decode attention backend
> 
> And if they are not provided, we will use attention backend.

OK, I will do it.