@Qiaolin-Yu 2025-05-16
> > Nice PR! And thanks for the detailed benchmark!
> > How about provide two optional server args:
> > 
> > * prefill attention backend
> > * decode attention backend
> > 
> > And if they are not provided, we will use attention backend.
> 
> OK, I will do it.

Hi @TianQiLin666666 , thanks for the great work! I will finish the follow-up on this PR. I have opened a new PR #6338, and added you as a coauthor 👀.